Actual CloudFormation stuff! Like working code!
We are going to use YAML for this project...wayyyyy easier to read I was shocked since I usually prefer brackets
Also you can add in comments in the templates...this is huge too
Also there is a lot of bloat in this document...I was messing around with the Designer which basically adds in its own version of CSS...not sure if I like or want to keep and am really not sure. Sorry it impacts readability a little bit.