Added in section for removing root user privileges
It's kind of straining my eyes looking at this. I'm going to print this off at home in a little bit and go through it and see if there is 1). anything we currently need to edit and 2). if there are any more sections we need to add (like I did with the root user) and 3). learn some stuff from what you've put up